requirements-dev.txt
# Minimal dev requirements for quick local debugging (no heavy extras)
requests==2.31.0
PyYAML==6.0.1
redis==5.0.0
pandas==2.0.3
numpy==1.24.3
pytest==7.4.2
pytest-cov==4.1.0
black==23.7.0
flake8==6.1.0
python-dotenv==1.0.0
matplotlib==3.7.2
